requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
faiss-cpu
pypdf
python-dotenv
sentence-transformers